feat(frontend): 글로벌 theme 변경
