--- frontend/src/styles/fonts.js
@@ -1,13 +1,0 @@
-import { css } from "@emotion/react";
-
-export const fontsStyle = css`
-  @import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/variable/pretendardvariable-dynamic-subset.css");
-
-  @font-face {
-    font-family: "BinggraeSamanco";
-    src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-10@1.0/BinggraeSamanco-Bold.woff")
-      format("woff");
-    font-weight: normal;
-    font-style: normal;
-  }
-`;
