+++ frontend/src/styles/theme/container.js
@@ -2,6 +2,6 @@ import { defineStyleConfig } from "@chakra-ui/react";
 
 export const CustomContainer = defineStyleConfig({
   baseStyle: {
-    paddingX: 10,
+    paddingX: "40px",
   },
 });
