+++ frontend/src/styles/theme/index.js
@@ -3,6 +3,13 @@ import { CustomButton } from "./button";
 import { CustomContainer } from "./container";
 
 export const theme = extendTheme({
+  styles: {
+    global: {
+      body: {
+        bgColor: "gray.50",
+      },
+    },
+  },
   fonts: {
     heading: `"Pretendard Variable", Pretendard, -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"`,
     body: `"Pretendard Variable", Pretendard, -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"`,
